// src/activity/types/activity.enum.ts
export enum ActivityEnum {
  LIKE_EARN = 'LIKE_EARN',
  LIKE_SPEND = 'LIKE_SPEND',
  IMAGE_GENERATE_SPEND = 'IMAGE_GENERATE_SPEND',
  VIDEO_GENERATE_SPEND = 'VIDEO_GENERATE_SPEND',
  CONTEST_CLOSE = 'CONTEST_CLOSE',
  CONTEST_OPEN = 'CONTEST_OPEN',
  CONTEST_WIN = 'CONTEST_WIN',
  DAILY_REWARD = 'DAILY_REWARD',
  SHARE_REWARD = 'SHARE_REWARD',
  ADMIN_REPORT = 'ADMIN_REPORT',
  ADMIN_CONTEST_REVIEW = 'ADMIN_CONTEST_REVIEW',
  ADMIN_REPORT_REVIEW = 'ADMIN_REPORT_REVIEW',
  ADMIN_CONTEST_WON = 'ADMIN_CONTEST_WON',
}